requirements.txt
aiohttp==3.11.10
astroid==3.3.6
async-timeout==5.0.1
attrs==24.2.0
cffi==1.17.1
chardet==5.2.0
colorama==0.4.6
cryptography==44.0.0
discord.py==2.4.0
ffmpeg==1.4
idna==3.10
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
multidict==6.1.0
pycparser==2.22
pylint==3.3.2
pymongo==4.10.1
PyNaCl==1.5.0
pyspnego==0.11.2
six==1.17.0
smbprotocol==1.15.0
toml==0.10.2
typing_extensions==4.12.2
wrapt==1.17.0
yarl==1.18.3
yt-dlp==2024.12.6
youtube-search-python==1.6.6
python-dotenv==1.0.1
audioop-lts==0.2.1
